Move gulp build to es6, split into task files
