+++ gulpfile.babel.js
@@ -1,0 +1,3 @@
+import req from 'require-dir'
+
+req('tasks')
