+++ tasks/build.js
@@ -1,0 +1,6 @@
+import gulp from 'gulp'
+import runSequence from 'run-sequence'
+
+gulp.task('build', function (cb) {
+  runSequence(['scripts', 'styles', 'static'], cb)
+})
