+++ tasks/default.js
@@ -1,0 +1,3 @@
+import gulp from 'gulp'
+
+gulp.task('default', ['dev'])
